third_party/blink/web_tests/fast/js/function-declarations-expected.txt

This page tests function declarations inside various block structures and control statements.

Before executing blocks containing function declarations: 
-----
PASS: 'f1' in window should be true and is.
PASS: 'f2' in window should be true and is.
PASS: 'f3' in window should be true and is.
PASS: 'f4' in window should be true and is.
PASS: 'f5' in window should be true and is.
PASS: 'f6' in window should be true and is.
PASS: 'f7' in window should be true and is.
PASS: 'f8' in window should be true and is.
PASS: 'f9' in window should be true and is.

After executing blocks containing function declarations: 
-----
PASS: 'f1' in window should be true and is.
PASS: 'f2' in window should be true and is.
PASS: 'f3' in window should be true and is.
PASS: 'f4' in window should be true and is.
PASS: 'f5' in window should be true and is.
PASS: 'f6' in window should be true and is.
PASS: 'f7' in window should be true and is.
PASS: 'f8' in window should be true and is.
PASS: 'f9' in window should be true and is.
